backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Nikita <yakuri2006@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.99.1"
requests = "^2.31.0"
aiohttp = "^3.8.4"
jpype1 = "^1.4.1"
cython = "^3.0.0"
postgis = "^1.0.4"
asyncpg = "^0.28.0"
pyproj = "^3.6.0"
geoalchemy2 = "^0.14.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.20"}


[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.22.0"}
ipykernel = "^6.25.1"
stubgenj = "^0.2.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
